Incorporate config for focusmaster patch
Applied by running:

    patch -p1 config.h < <(git show patch-focusmaster config.def.h)

It didn't apply cleanly, but I looked at the .rej file produced and
applied it easily by editing the source directly.

patch-focusmaster is the name of the branch that contains the patch.
When referred to in a `git show` command like this it's as if we'd
entered the SHA for the latest commit on the patch-focusmaster branch.

What's happening? `git show` shows the change that a patch introduced to
config.def.h (the default config). The `patch` command then applies that
change to config.h.