Incorporate config for scratchpad patch
Applied by running:

    patch -p1 config.h < <(git show <sha> config.def.h)

What's happening? `git show` shows the change that a patch introduced to
config.def.h (the default config). The `patch` command then applies that
change to config.h.

<sha> was actually the commit that contained the change to config.def.h
that comes with the scratchpad patch.

I usually apply these config changes by referencing the name of the
branch that contains the patch (e.g. patch-scratchpad), but in this case
I knew the commit on the patch-scratchpad branch wouldn't apply cleanly.

I'd already had to resolve a conflict with another patch that I'm using
on the selected-patches branch, so I just referenced that commit to get
a version of the change that I knew would apply to the config changes
that have already been made.