Apply the focusmaster patch
There's no patch page on suckless.org for this patch yet, so I
downloaded it using these two commands:

    patch -p1 < <(curl -s https://raw.githubusercontent.com/jceb/dwm-patches/master/patches/configh_focusmaster.patch)
    git apply <(curl -s https://raw.githubusercontent.com/jceb/dwm-patches/master/patches/focusmaster.patch)

The patch to config.def.h had to be run with `patch` rather than `git
apply` as it wouldn't apply cleanly and needed editing manually. This is
because the repo I downloaded it from stores it as a patch against the
owner's customised version of dwm.